Refactor Dockerfile and add CRUD functionality
Refactor the Dockerfile to improve readability and add CRUD functionality to the application. This includes creating, updating, and deleting items using HTTP methods. Also, update the go.mod file to include the required dependency for the CRUD functionality. Finally, add a Kubernetes deployment and service configuration file to deploy the application in a Kubernetes cluster.